1.5 - Improved Sensitivity Value, Filtering, Error Testing
Made it so sensitivity value changes the canny thresholds. Added error checking if no screws are detected, putting -1 in all fields of a single entry on scewLocations global. Removed magic numbers from additional filter size filtering, which still needs to be incorporated with GUI using a global. Added line to save output to a file for better viewing and later GUI use.